Add files via upload
Fixed some major bugs concerning the set up of unseen titles and saving of genre scores into the database which occurred during the further modularisation of the code. This has stopped titles being repeatedly saved to the unseen titles in the database. I have changed the user._genrescors to a dictionary rather than a list to accomodate these changes, and for fasting querying of genres that have associated 'average ratings'.